Move some of the helpers to the proper modules.
